Move persistence services one folder level up
