--- src/new_model/runtime/persistence/flow_node_instance/index.ts
@@ -1,1 +1,0 @@
-export * from './service';
