+++ src/new_model/runtime/persistence/index.ts
@@ -1,2 +1,2 @@
-export * from './flow_node_instance/index';
-export * from './process_model/index';
+export * from './flow_node_instance_service';
+export * from './process_model_service';
