--- src/new_model/runtime/persistence/process_model/index.ts
@@ -1,1 +1,0 @@
-export * from './service';
